:sparkles: feat: 사용자 가입 조직 조회 기능 구현 (#28)
