--- src/main/java/com/notitime/noffice/domain/OrganizationCategory.java
@@ -1,28 +1,0 @@
-package com.notitime.noffice.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class OrganizationCategory {
-
-	@Id
-	@GeneratedValue
-	private Long id;
-
-	@ManyToOne
-	@JoinColumn(name = "organization_id")
-	private Organization organization;
-
-	@ManyToOne
-	@JoinColumn(name = "category_id")
-	private Category category;
-}
