+++ src/main/java/com/notitime/noffice/domain/organization/model/OrganizationCategory.java
@@ -1,10 +1,7 @@
-package com.notitime.noffice.domain;
+package com.notitime.noffice.domain.organization.model;
 
-import com.notitime.noffice.domain.member.model.Member;
+import com.notitime.noffice.domain.Category;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
@@ -16,20 +13,17 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class OrganizationMember {
+public class OrganizationCategory {
 
 	@Id
 	@GeneratedValue
 	private Long id;
 
-	@Enumerated(EnumType.STRING)
-	private OrganizationRole role;
-
-	@ManyToOne(fetch = FetchType.LAZY)
+	@ManyToOne
 	@JoinColumn(name = "organization_id")
 	private Organization organization;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id")
-	private Member member;
+	@ManyToOne
+	@JoinColumn(name = "category_id")
+	private Category category;
 }
