+++ src/main/java/com/notitime/noffice/domain/organization/model/OrganizationMember.java
@@ -1,13 +1,15 @@
-package com.notitime.noffice.domain;
+package com.notitime.noffice.domain.organization.model;
 
+import com.notitime.noffice.domain.OrganizationRole;
 import com.notitime.noffice.domain.member.model.Member;
-import com.notitime.noffice.domain.organization.model.Organization;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,32 +17,20 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class Announcement extends BaseTimeEntity {
+public class OrganizationMember {
 
 	@Id
+	@GeneratedValue
 	private Long id;
 
-	private String title;
-
-	private String content;
-
-	private String coverImage;
-
-	private String place_link_title;
-
-	private String place_link_url;
-
-	private LocalDateTime startAt;
-
-	private LocalDateTime endAt;
-
-	private LocalDateTime noticeAt;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "creator_id")
-	private Member member;
+	@Enumerated(EnumType.STRING)
+	private OrganizationRole role;
 
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "organization_id")
 	private Organization organization;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id")
+	private Member member;
 }
