+++ src/main/java/com/notitime/noffice/response/OrganizationResponses.java
@@ -1,6 +1,6 @@
 package com.notitime.noffice.response;
 
-import com.notitime.noffice.domain.OrganizationMember;
+import com.notitime.noffice.domain.organization.model.OrganizationMember;
 import java.util.List;
 
 public record OrganizationResponses(List<OrganizationResponse> organizations) {
